+++ pom.xml
@@ -117,12 +117,6 @@
       <version>${jmh.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.reaktivity</groupId>
-      <artifactId>nukleus-tcp</artifactId>
-      <version>0.3</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
