+++ src/main/java/org/reaktivity/nukleus/http2/internal/types/stream/Http2DataFW.java
@@ -75,7 +75,6 @@ public class Http2DataFW extends Flyweight
 
     public int streamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         return buffer().getInt(offset() + STREAM_ID_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
