+++ src/main/java/org/reaktivity/nukleus/http2/internal/types/stream/Http2FrameFW.java
@@ -71,7 +71,6 @@ public final class Http2FrameFW extends Flyweight
 
     public int streamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         return buffer().getInt(offset() + STREAM_ID_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
