+++ src/main/java/org/reaktivity/nukleus/http2/internal/types/stream/Http2GoawayFW.java
@@ -83,13 +83,11 @@ public class Http2GoawayFW extends Flyweight
     // streamId == 0, caller to validate
     public int streamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         return buffer().getInt(offset() + STREAM_ID_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
 
     public int lastStreamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         int streamId = buffer().getInt(offset() + STREAM_ID_OFFSET) & 0x7F_FF_FF_FF;
         assert streamId == 0;
         return streamId;
