+++ src/main/java/org/reaktivity/nukleus/http2/internal/types/stream/Http2RstStreamFW.java
@@ -76,11 +76,10 @@ public class Http2RstStreamFW extends Flyweight
     // streamId != 0, caller to validate
     public int streamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         return buffer().getInt(offset() + STREAM_ID_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
 
-    public long errorCode()
+    public int errorCode()
     {
         return buffer().getInt(offset() + PAYLOAD_OFFSET, BIG_ENDIAN);
     }
