+++ src/main/java/org/reaktivity/nukleus/http2/internal/types/stream/Http2WindowUpdateFW.java
@@ -69,13 +69,11 @@ public class Http2WindowUpdateFW extends Flyweight
 
     public int streamId()
     {
-        // Most significant bit is reserved and is ignored when receiving
         return buffer().getInt(offset() + STREAM_ID_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
 
     public int size()
     {
-        // Most significant bit is reserved
         return buffer().getInt(offset() + PAYLOAD_OFFSET, BIG_ENDIAN) & 0x7F_FF_FF_FF;
     }
