+++ src/test/java/org/reaktivity/nukleus/http2/internal/types/stream/HpackHeaderBlockFWTest.java
@@ -143,10 +143,10 @@ public class HpackHeaderBlockFWTest
 
         HpackHeaderBlockFW fw = new HpackHeaderBlockFW.Builder()
                 .wrap(buf, 1, buf.capacity())
-                .header(hf -> hf.indexed(2))      // :method: GET
-                .header(hf -> hf.indexed(6))      // :scheme: http
-                .header(hf -> hf.indexed(4))      // :path: /
-                .header(hf -> hf.literal(l -> l.type(INCREMENTAL_INDEXING).name(1).value("www.example.com")))
+                .header(h -> h.indexed(2))      // :method: GET
+                .header(h -> h.indexed(6))      // :scheme: http
+                .header(h -> h.indexed(4))      // :path: /
+                .header(h -> h.literal(l -> l.type(INCREMENTAL_INDEXING).name(1).value("www.example.com")))
                 .build();
 
         assertEquals(21, fw.limit());
@@ -169,11 +169,11 @@ public class HpackHeaderBlockFWTest
 
         HpackHeaderBlockFW fw = new HpackHeaderBlockFW.Builder()
                 .wrap(buf, 1, buf.capacity())
-                .header(hf -> hf.indexed(2))        // :method: GET
-                .header(hf -> hf.indexed(6))        // :scheme: http
-                .header(hf -> hf.indexed(4))        // :path: /
-                .header(hf -> hf.indexed(62))       // :authority: www.example.com
-                .header(hf -> hf.literal(l -> l.type(INCREMENTAL_INDEXING).name(24).value("no-cache")))
+                .header(h -> h.indexed(2))        // :method: GET
+                .header(h -> h.indexed(6))        // :scheme: http
+                .header(h -> h.indexed(4))        // :path: /
+                .header(h -> h.indexed(62))       // :authority: www.example.com
+                .header(h -> h.literal(l -> l.type(INCREMENTAL_INDEXING).name(24).value("no-cache")))
                 .build();
         assertEquals(15, fw.limit());
 
@@ -196,11 +196,11 @@ public class HpackHeaderBlockFWTest
 
         HpackHeaderBlockFW fw = new HpackHeaderBlockFW.Builder()
                 .wrap(buf, 1, buf.capacity())
-                .header(y -> y.indexed(2))    // :method: GET
-                .header(y -> y.indexed(7))    // :scheme: https
-                .header(y -> y.indexed(5))    // :path: /index.html
-                .header(y -> y.indexed(63))   // :authority: www.example.com
-                .header(y -> y.literal(z -> z.type(INCREMENTAL_INDEXING).name("custom-key").value("custom-value")))
+                .header(h -> h.indexed(2))    // :method: GET
+                .header(h -> h.indexed(7))    // :scheme: https
+                .header(h -> h.indexed(5))    // :path: /index.html
+                .header(h -> h.indexed(63))   // :authority: www.example.com
+                .header(h -> h.literal(l -> l.type(INCREMENTAL_INDEXING).name("custom-key").value("custom-value")))
                 .build();
         assertEquals(30, fw.limit());
